make ldtpme being installed with setup.py
