+++ setup.py
@@ -23,13 +23,13 @@ Headers in this file shall remain intact.
 from distutils.core import setup
 
 setup(name="ldtp",
-      version="3.5.0",
+      version="3.5.1",
       description="Linux Desktop Testing Project Version 2",
       maintainer="Nagappan Alagappan",
       maintainer_email="nagappan@gmail.com",
       url="http://ldtp.freesktop.org",
       license="GNU Lesser General Public License (LGPL)",
-      packages=["ldtp", "ldtpd", "ooldtp", "ldtputils"],
+      packages=["ldtp", "ldtpd", "ooldtp", "ldtputils", "ldtpme"],
       long_description="Linux Desktop Testing Project is aimed at producing " \
           "high quality cross platform GUI test automation framework and cutting-edge tools that " \
           "can be used to test GNU/Linux/Windows/Mac Desktop and improve it. It uses the " \
